transform.api/pyproject.toml
[project]
name = "transform.api"
version = "1.2.1"
description = "资源转换服务"
license = "MIT"
authors = [
    { name = "menglb", email = "lingbinmeng@hotmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "pypdfium2 (>=4.30.1,<5.0.0)",
    "grpcio (>=1.72.1,<2.0.0)",
    "grpcio-tools (>=1.72.1,<2.0.0)",
    "spacy>=3.8.7",
    "en-core-web-sm @ https://lib.lc.jtxuexi.com/en_core_web_sm-3.8.0-py3-none-any.whl",
    "zh-core-web-lg @ https://lib.lc.jtxuexi.com/zh_core_web_lg-3.8.0-py3-none-any.whl",
    "pip>=25.1.1",
    "python-docx>=1.2.0",
    "bs4>=0.0.2",
    "pandas>=2.3.0",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "openai>=2.9.0",
    "xinference-client>=1.14.0",
]
[tool.uv]
allow-insecure-host = ["lib.lc.jtxuexi.com"]
